Adds: CSV processing and first method of second iteration
