+++ RakeFile
@@ -1,0 +1,7 @@
+require 'rake/testtask'
+
+Rake::TestTask.new do |t|
+    t.pattern = "test/**/*_test.rb"
+end
+
+task default: ["test"]
