+++ test/test_helper.rb
@@ -1,0 +1,7 @@
+require 'simplecov'
+SimpleCov.start
+require 'minitest/autorun'
+require 'minitest/pride'
+require './lib/stat_tracker'
+require 'csv'
+require 'pry'
